feat(handdecomp): port StartMuralBackgroundAlt → matching C (screen NEAR sign-ext)
u8 flag sign-ext (lsr->asr); int param + (s8) cast — works here since flag
is used in a single simple comparison (no reg-alloc restructure). mc 7658->7659.

+++ include/functions.h
@@ -371,7 +371,7 @@ void DrawStatBarGfx(
 // ??? BackgroundSlide_Init(???);
 // ??? BackgroundSlide_Loop(???);
 ProcPtr StartMuralBackground(ProcPtr, void*, int);
-ProcPtr StartMuralBackgroundAlt(ProcPtr parent, void * vram, int pal, u8 flag);
+ProcPtr StartMuralBackgroundAlt(ProcPtr parent, void * vram, int pal, int flag);
 ProcPtr StartMuralBackgroundExt(ProcPtr, void*, int, int, int);
 void EndMuralBackground(void);
